fix: a literal's width, and two constructs no backend but Nim had built
Writing core.num and core.hash in Tuck — the first stdlib modules that are
not thin shims — turned up six defects, five of them invisible because Nim
happened to work.

**A literal's width, through a `?T` wrapper.** `return 0` in an `-> i64?` fn
is wrapped by codegen (`tok(0)`), so the width the literal must carry is the
PAYLOAD's. synthLit reads the expected-type channel but only recognised a
BARE numeric name; `?i64` is a tkApp, so the hint was dropped and the literal
stayed `int`. Nim refused `TuckResult[int]` for `TuckResult[int64]`, Odin
refused `TuckResult($T=int)` for `TuckResult($T=i64)`. The channel is
unwrapped now.

**A match arm inherited the SUBJECT's width.** The channel is set to the
match subject so a bare variant in an arm body resolves — it was called
`expectedVariantType` before it was generalised. Offering it for anything
else is a hint the body never asked for: `match r.err: Empty: 42 sys::exit`
has a u16 subject, and the arm emitted `exit(42'u16)`, which Nim's
`exit(int)` refuses. Only a SUM subject is offered now (`variantHint`).

**Odin had no literal-width treatment at all.** A bare Odin literal is `int`
(or `f64`), and Odin converts to a fixed-width type only explicitly, so a
literal in an inferred position now spells `i64(0)` — the same job Nim's
`'u64` suffix and D's `L` already did.

**D constructed where it had to cast.** `{value: n} u8` emitted `ubyte(x)`,
which is a type constructor performing only the conversions D would do
implicitly: "cannot implicitly convert expression of type ulong to ubyte".
A Tuck conversion call is explicit by construction, so it is a `cast` — for
every numeric width and bool, never for `string`.

**Odin lost `push` the moment a module imported std/seq.** The extern
forwarder marked its type param with a SECOND copy of the rule, testing for
a bare named `T` — which `Seq[T]` (rendered `[dynamic]T`) is not. The sigil
landed on the second parameter and the first named an undeclared `T`. It now
calls markFirstTypeParam, the same proc fnParamList uses; the duplicate rule
is gone.

**`match r.err` compiled on Nim alone.** Odin emitted `default:` — not a
keyword there, so it parsed as a variable declaration — and D emitted
`case E.Empty:` against a u16 subject, "undefined identifier `E`". Odin gets
a bare `case:`, D gets the folded error code and a plain switch (D's `final
switch` is enum-only, and a plain switch then requires a default).

Also `nimPrimitive` exported, and the bit operations added to all three
runtimes for std/bits.

Suite green including the ratchet. The examples diff is width-only: `1L` in
D, `u64(1000)` in Odin, `1000'u64` and `1'u32` in Nim — every one of them a
literal now saying what the checker already knew it was.

+++ examples/01-data-flow.odin
@@ -60,7 +60,7 @@ tuck_playTrack :: proc(payload: $T) {
 
 
 tuck_main :: proc () {
-  tuck_request := TRec_url_timeout_E49C{url = "example.com", timeout = time.tuck_ms(5)}
+  tuck_request := TRec_url_timeout_E49C{url = "example.com", timeout = time.tuck_ms(u32(5))}
   tuck_response := tuck_selectEpisodes(tuck_parse(tuck_fetch(tuck_request)))
   tuck_feed := tuck_fetch("https://example.com/feed")
   if tuck_feed.hasNew {
