+++ examples/34-ffi-cstring.odin
@@ -19,7 +19,7 @@ foreign z {
 tuck_main :: proc () {
   tuck_v := zlibVersion()
   console.printLine(tuck_v)
-  tuck_b := compressBound(1000)
+  tuck_b := compressBound(u64(1000))
   if (tuck_b == 1013) {
       sys.exit(0)
   }
